==> code/step4.ipynb <==
import pandas as pd
import statsmodels.api as sm

T = [2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021]
for t in T:
    df = pd.read_csv('../data/step3/{}.csv'.format(t), index_col=0)
    FX = df[['rm-rf', 'SMB', 'HML', 'RMW', 'CMA']]
    FY = df[['ri300-rf']]
    FX = sm.add_constant(FX)

    model = sm.OLS(FY, FX).fit()
    
    print(model.summary(), '\n\n')
